[WIP] StreamSrc/Sink test passing
