--- pymtl3/stdlib/dstruct/ifcs/__init__.py
@@ -1,1 +1,0 @@
-from .ifcs import EnqIfc, DeqIfc, RecvIfc, SendIfc
