+++ pymtl3/stdlib/test_utils/__init__.py
@@ -6,6 +6,3 @@ from .test_helpers import (
     run_sim,
     run_test_vector_sim,
 )
-from .test_masters import TestMasterCL
-from .test_sinks import TestSinkFL
-from .test_srcs import TestSourceFL
